moving endpoint over to samples
