+++ src/backend/aspen/main.py
@@ -9,4 +9,3 @@ from aspen.app.views.phylo_trees import auspice_view, phylo_trees  # noqa: F401
 from aspen.app.views.sample import samples  # noqa: F401
 from aspen.app.views.usergroup import usergroup  # noqa: F401
 from aspen.app.views.usher import get_usher_tree_options  # noqa: F401
-from aspen.app.views.validation import validate_ids  # noqa: F401
